Extract credit card scripts into standalone JS file
