+++ resources/views/portal/ninja2020/gateways/mollie/credit_card/pay.blade.php
@@ -2,8 +2,6 @@
 ctrans('texts.credit_card')])
 
 @section('gateway_head')
-    @dd($contact->preferredLocale())
-
     <meta name="mollie-testmode" content="{{ $gateway->company_gateway->getConfigField('testMode') }}">
     <meta name="mollie-publicId" content="{{ $gateway->company_gateway->getConfigField('publicId') }}">
 
@@ -83,159 +81,10 @@ ctrans('texts.credit_card')])
     @endcomponent
 
     @include('portal.ninja2020.gateways.includes.save_card')
-
     @include('portal.ninja2020.gateways.includes.pay_now')
 @endsection
 
 @section('gateway_footer')
     <script src="https://js.mollie.com/v1/mollie.js"></script>
-
-    <script>
-        class _Mollie {
-            constructor() {
-                this.mollie = Mollie(document.querySelector('meta[name=mollie-publicId]')?.content, {
-                    testmode: document.querySelector('meta[name=mollie-testmode]')?.content,
-                });
-            }
-
-            createCardHolderInput() {
-                let cardHolder = this.mollie.createComponent("cardHolder");
-                cardHolder.mount("#card-holder");
-
-                let cardHolderError = document.getElementById("card-holder-error");
-
-                cardHolder.addEventListener("change", function(event) {
-                    if (event.error && event.touched) {
-                        cardHolderError.textContent = event.error;
-                    } else {
-                        cardHolderError.textContent = "";
-                    }
-                });
-
-                return this;
-            }
-
-            createCardNumberInput() {
-                let cardNumber = this.mollie.createComponent("cardNumber");
-                cardNumber.mount("#card-number");
-
-                let cardNumberError = document.getElementById("card-number-error");
-
-                cardNumber.addEventListener("change", function(event) {
-                    if (event.error && event.touched) {
-                        cardNumberError.textContent = event.error;
-                    } else {
-                        cardNumberError.textContent = "";
-                    }
-                });
-
-                return this;
-            }
-
-            createExpiryDateInput() {
-                let expiryDate = this.mollie.createComponent("expiryDate");
-                expiryDate.mount("#expiry-date");
-
-                let expiryDateError = document.getElementById("expiry-date-error");
-
-                expiryDate.addEventListener("change", function(event) {
-                    if (event.error && event.touched) {
-                        expiryDateError.textContent = event.error;
-                    } else {
-                        expiryDateError.textContent = "";
-                    }
-                });
-
-                return this;
-            }
-
-            createCvvInput() {
-                let verificationCode = this.mollie.createComponent("verificationCode");
-                verificationCode.mount("#cvv");
-
-                let verificationCodeError = document.getElementById(
-                    "cvv-error"
-                );
-
-                verificationCode.addEventListener("change", function(event) {
-                    if (event.error && event.touched) {
-                        verificationCodeError.textContent = event.error;
-                    } else {
-                        verificationCodeError.textContent = "";
-                    }
-                });
-
-                return this;
-            }
-
-            handlePayNowButton() {
-                document.getElementById('pay-now').disabled = true;
-
-                if (document.querySelector('input[name=token]').value !== '') {
-                    document.querySelector('input[name=gateway_response]').value = '';
-
-                    return document.getElementById('server-response').submit();
-                }
-
-                this.mollie.createToken().then(function(result) {
-                    let token = result.token;
-                    let error = result.error;
-
-                    if (error) {
-                        document.getElementById('pay-now').disabled = false;
-
-                        let errorsContainer = document.getElementById('errors');
-                        errorsContainer.innerText = error.message;
-                        errorsContainer.hidden = false;
-
-                        return;
-                    }
-
-                    let tokenBillingCheckbox = document.querySelector(
-                        'input[name="token-billing-checkbox"]:checked'
-                    );
-
-                    if (tokenBillingCheckbox) {
-                        document.querySelector('input[name="store_card"]').value =
-                            tokenBillingCheckbox.value;
-                    }
-
-                    document.querySelector('input[name=gateway_response]').value = token;
-                    document.querySelector('input[name=token]').value = '';
-
-                    document.getElementById('server-response').submit();
-                });
-            }
-
-            handle() {
-                this
-                    .createCardHolderInput()
-                    .createCardNumberInput()
-                    .createExpiryDateInput()
-                    .createCvvInput();
-
-                Array
-                    .from(document.getElementsByClassName('toggle-payment-with-token'))
-                    .forEach((element) => element.addEventListener('click', (element) => {
-                        document.getElementById('mollie--payment-container').classList.add('hidden');
-                        document.getElementById('save-card--container').style.display = 'none';
-                        document.querySelector('input[name=token]').value = element.target.dataset.token;
-                    }));
-
-                document
-                    .getElementById('toggle-payment-with-credit-card')
-                    .addEventListener('click', (element) => {
-                        document.getElementById('mollie--payment-container').classList.remove('hidden');
-                        document.getElementById('save-card--container').style.display = 'grid';
-                        document.querySelector('input[name=token]').value = "";
-                    });
-
-                document
-                    .getElementById('pay-now')
-                    .addEventListener('click', () => this.handlePayNowButton());
-            }
-        }
-
-        new _Mollie().handle();
-    </script>
+    <script src="{{ asset('js/clients/payments/mollie-credit-card.js') }}"></script>
 @endsection
